version 1.0.0 fixes and layout update
invalid cache fix coming from windows.
removal of useless buttons.
removal of functions from useless options.
removal of graphics and useless items.
.conf file correction
adding nodes and whitelist in .conf file
addition of standard coin gates

the version does not change, as it is not a version update, but a version correction.